Refactored benchmarking suite

+++ include/simdmonte/pricer/params.h
@@ -20,7 +20,7 @@ struct Params {
   params::NormalMethod normal_method = params::NormalMethod::BoxMuller;
   // default constructor only here for use in gtest -- should find another way
   Params() {};
-  Params(int n_steps_, int n_sims_, params::UnderlyingModel underlying_model_, params::NormalMethod normal_method_) :
+  Params(int n_steps_, long n_sims_, params::UnderlyingModel underlying_model_, params::NormalMethod normal_method_) :
     n_steps(n_steps_), n_sims(n_sims_), underlying_model(underlying_model_), normal_method(normal_method_) {};
 };
